% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dfHeatMap.R
\name{dfHeatMap}
\alias{dfHeatMap}
\title{dfHeatMap}
\usage{
dfHeatMap(dat, myTitle, lowLim = 0, highLim = 1)
}
\arguments{
\item{dat}{the data frame to be plotted}

\item{lowLim}{the lower limit for the values}

\item{highLim}{the upperr limit for the values}

\item{dat}{myTitle a title for the plot}

\item{dat}{Input data frame}
}
\value{
zp1 The ggplot2 plot
}
\description{
Plot a heat map from a data frame using ggplot, passing limits
}
\keyword{keywords}

