% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multiplotGG2.R
\name{multiplotGG}
\alias{multiplotGG}
\title{multiplotGG2}
\usage{
multiplotGG(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{an unamed list of ggplot2 plots}

\item{plotlist}{a named list of ggplot2 plots}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.

If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.}
}
\description{
Plot multiple ggplot2 plots onto a grid
Copied into a package (shamelessly) from
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
to implement DRY. Note: removed the unused file parameter
}
\keyword{keywords}

