% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sortAscendingDataFrameCols.R
\name{sortAscendingDataFrameCols}
\alias{sortAscendingDataFrameCols}
\title{sortAscendingDataFrameCols}
\usage{
sortAscendingDataFrameCols(dat)
}
\arguments{
\item{dat}{Input data frame}
}
\value{
The sorted data frame
}
\description{
Sort each column of a data frame into ascending order. Useful for
a few features can be flipped.
}
\keyword{keywords}

